Update board and cells to track clicks
